task 3 done

+++ file_io/3-cp.c
@@ -58,7 +58,7 @@ void copy_content(int fd_from, int fd_to,
 		if (n_written != n_read)
 			print_error(99, "Error: Can't write to %s\n", file_to);
 	}
-	if (n_read == (size_t)-1)
+	if ((int)n_read == -1)
 		print_error(98, "Error: Can't read from file %s\n", file_from);
 }
